+++ SK.WebApp/src/services/languages.ts
@@ -1,0 +1,13 @@
+export enum anyLangageCodeKeys {
+  polish = 'POLISH',
+  english = 'ENGLISH',
+}
+
+export type languageCodeMap = {
+  [key in anyLangageCodeKeys]: string;
+};
+
+export const anyLanguageCode: languageCodeMap = {
+  POLISH: 'pl-PL',
+  ENGLISH: 'en-US',
+};
