+++ SK.WebApp/src/services/utils.ts
@@ -1,10 +1,5 @@
-import {
-  apiAddress,
-  anyEndpoint,
-  anyLanguageCode,
-  anyEndpointKeys,
-  anyLangageCodeKeys,
-} from './consts';
+import { apiAddress, anyEndpointKeys, anyEndpoint } from './endpoints';
+import { anyLanguageCode, anyLangageCodeKeys } from './languages';
 
 interface getEndpointWithLanguageCodeArgs {
   endpointKey: anyEndpointKeys;
